requirements.txt
aiohttp==3.8.3
black==22.12.0
certifi==2022.12.7
demoji==1.1.0
discord.py==2.1.0
fuzzywuzzy==0.18.0
<<<<<<< HEAD
=======
gitdb==4.0.10
GitPython==3.1.30
isort==5.12.0
idna==3.4
Levenshtein==0.20.9
multidict==6.0.4
mypy-extensions==0.4.3
>>>>>>> origin/public-refactor
numpy==1.24.1
pymongo==4.3.3
python-dotenv==0.21.0
requests==2.28.1
rich==13.0.0